Moved timer functions to their own module
